requirements.txt
PyMuPDF
google-generativeai
python-dotenv
Flask
markdown2
